refactor: extract executives/roles router (PR-8, Phase 1)
